requirements.txt
# NOTE: in order to install the requirements, please use the following commands
# python -m venv env
# activate your venv!
# pip install --find-links https://download.pytorch.org/whl/torch_stable.html -r requirements.txt
ipykernel
numpy>=1.18.1
pandas==1.0.5
matplotlib
azureml-core
azureml-dataprep[pandas,fuse]==2.0.7
azureml-sdk>=1.1.5

"""
azure-cosmos==3.1.2
azureml-sdk>=1.1.5
azureml-dataprep[pandas,fuse]==2.0.7
#azure-ai-textanalytics==1.0.0b3 # TO BE UPGRADED -> 5.0.0
gensim==3.8.0
spacy==2.3.2 #UPGRADED
transformers==3.0.2 #UPGRADED
farm==0.4.7 #UPGRADED
flair==0.6 #UPGRADED
selenium==3.141.0
bs4
boto3
scipy>=1.3.2
sklearn
seqeval
dotmap==1.3.0
pyyml
opencensus==0.7
opencensus-ext-azure==1
fsspec
##DEPLOY ONLY
azure-keyvault==1.1.0
azure-identity==1.3.1
azure-keyvault-secrets==4.1.0
##LOCAL ONLY
pylint>=2.4.4
pytest>=5.3.5
flake8>=3.7.9
ipykernel>=5.1.4
streamlit==0.65
tqdm
torch==1.5.1+cpu 
torchvision==0.6.1+cpu
"""